requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
djangorestframework==3.15.0
drf-spectacular==0.28.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
swagger_ui_bundle==1.1.0
uritemplate==4.2.0
validate_docbr==1.11.1
